+++ app/components/StarRating.js
@@ -4,22 +4,20 @@ import { Image, Text } from 'react-native'
 const StarRating = props => {
 	return (
 		<>
-			{[...Array(props.card.rating)].map(() => (
-				<>
-					<Image
-						className='mx-[5px] h-[25px] w-[25px]'
-						source={require('../../assets/images/star.png')}
-					/>
-				</>
+			{[...Array(props.card.rating)].map((val, index) => (
+				<Image
+					key={index}
+					className='mx-[5px] h-[25px] w-[25px]'
+					source={require('../../assets/images/star.png')}
+				/>
 			))}
 			{props.card.rating < 5 &&
-				[...Array(5 - props.card.rating)].map(() => (
-					<>
-						<Image
-							className='mx-[5px] h-[25px] w-[25px]'
-							source={require('../../assets/images/star_empty.png')}
-						/>
-					</>
+				[...Array(5 - props.card.rating)].map((val, index) => (
+					<Image
+						key={index}
+						className='mx-[5px] h-[25px] w-[25px]'
+						source={require('../../assets/images/star_empty.png')}
+					/>
 				))}
 			{props.card.reviews && <Text>({props.card.reviews})</Text>}
 		</>
